+++ packages/mf-host/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+# [2.1.0](https://github.com/VirtoCommerce/vc-shell/compare/v2.0.11...v2.1.0) (2026-06-22)
+
+**Note:** Version bump only for package @vc-shell/mf-host
+
+## [2.0.11](https://github.com/VirtoCommerce/vc-shell/compare/v2.0.10...v2.0.11) (2026-06-22)
+
+**Note:** Version bump only for package @vc-shell/mf-host
+
 ## [2.0.10](https://github.com/VirtoCommerce/vc-shell/compare/v2.0.9...v2.0.10) (2026-06-15)
 
 **Note:** Version bump only for package @vc-shell/mf-host